cast an object.size() to (int)
 - Seems like sizes in R are generally size_t = unsigned long,
   but it's not universal. When I switched things to size_t, I got a
   complaint about one case where the size was plain long.

 - Looking more closely, it seems that:
             IntegerVector.size() is unsigned long
      *but*  NumericVector.size() is long

 - Sticking with signed ints for now, but should look into switching
   to proper size_t where I can.

+++ src/arrange_genes.cpp
@@ -30,7 +30,7 @@ IntegerVector arrange_genes(const NumericVector& start,
     result[0] = 1;
 
     for(int i=1; i<n; i++) {
-        for(int j=0; j<maxx.size(); j++) {
+        for(int j=0; j<(int)maxx.size(); j++) {
             if(start[i] > maxx[j]) {
                 result[i] = j+1;
                 maxx[j] = end[i];
